consolidate into single esxi server class
