+++ vm_automation/__init__.py
@@ -1,3 +1,2 @@
 from .esxiVm import esxiServer, esxiVm
-from .esxiSsh import esxiSsh
 from .workstationVm import workstationServer, workstationVm
